+++ frontend/src/knowledgePaths/KnowledgePathDetail.jsx
@@ -46,6 +46,11 @@ import BookmarkButton from '../bookmarks/BookmarkButton';
 import KnowledgePathDetailSkeleton from '../components/KnowledgePathDetailSkeleton';
 import BookClubReturnLink from '../bookClubs/BookClubReturnLink';
 import CryptoPaymentModal from '../events/CryptoPaymentModal';
+import ProductPaymentCheckout from '../payments/ProductPaymentCheckout';
+import {
+  createPathPurchaseBchPayment,
+  verifyPathPurchaseBchPayment,
+} from '../api/paymentsApi';
 
 // TODO: Add a progress bar to the knowledge path detail page
 const KnowledgePathDetail = () => {
@@ -926,15 +931,31 @@ const KnowledgePathDetail = () => {
         </Paper>
       )}
 
-      <CryptoPaymentModal
-        open={showPaymentModal}
-        onClose={() => setShowPaymentModal(false)}
-        pathPurchaseId={pathPurchaseId}
-        title={knowledgePath?.title}
-        priceUsd={knowledgePath?.reference_price}
-        productLabel="camino"
-        onPaymentComplete={handlePaymentComplete}
-      />
+      {knowledgePath?.bch_direct_available ? (
+        <ProductPaymentCheckout
+          open={showPaymentModal}
+          onClose={() => setShowPaymentModal(false)}
+          title={knowledgePath?.title}
+          priceUsd={knowledgePath?.reference_price}
+          productLabel="camino"
+          offerNowpayments
+          offerBch
+          createBchPayment={() => createPathPurchaseBchPayment(pathPurchaseId)}
+          verifyBchPayment={() => verifyPathPurchaseBchPayment(pathPurchaseId)}
+          nowpaymentsProps={{ pathPurchaseId }}
+          onPaid={handlePaymentComplete}
+        />
+      ) : (
+        <CryptoPaymentModal
+          open={showPaymentModal}
+          onClose={() => setShowPaymentModal(false)}
+          pathPurchaseId={pathPurchaseId}
+          title={knowledgePath?.title}
+          priceUsd={knowledgePath?.reference_price}
+          productLabel="camino"
+          onPaymentComplete={handlePaymentComplete}
+        />
+      )}
     </Container>
   );
 };